Added demo version of new interface that allows the same code to export destinations and transformations
