+++ README.md
@@ -1,24 +1,29 @@
 # Jitsu SDK (alpha)
 
-This is an alpha version of [Jitsu](https://github.com/jitsucom/jitsu) SDK. It contains a few packages that solves
-different challenges with Jitsu extensibility
+This is an alpha version of [Jitsu](https://github.com/jitsucom/jitsu) extension SDK. 
+It contains a few packages that build, test and execute jitsu extensions
 
-* **Jitsu CLI** — a CLI interface for developing jitsu plugins, published as an [npm package](https://www.npmjs.com/package/@jitsu/jitsu-cli). At the moment,
-only destination development works (`jitsu destination build`, `jitsu destination test` and soon `npx jitsu destination create`)
-* **Jitsu Tracking SDK** (not moved here yet, see [main repo](https://github.com/jitsucom/jitsu/tree/master/javascript-sdk))
+* **Jitsu CLI** — a CLI interface for developing jitsu extensions, published as an [npm package](https://www.npmjs.com/package/@jitsu/jitsu-cli). 
+See run `jitsu extension build`, `jitsu extension build` and soon `npx @jitsu/cli extension build`)
 
 ## Implementing Jitsu destination
 
 ```shell
-npx jitsu destination create
+npx jitsu extension create --type destination
 ```
 
-Will create an empty project. Then run:
+```shell
+npx jitsu extension create --type transformation
+```
 
-* `yarn build` to build a destination
+Will create an empty project. Depending on `--type`, it will be either `destination` extension, or `transformation` extension
+
+Run:
+
+* `yarn build` to build an extension
 * `yarn test` to run tests
 
-`yarn build` and `yarn test` are just wrappers around `jitsu destination build` and `jitsu destination test`
+`yarn build` and `yarn test` are just wrappers around `jitsu extension build` and `jitsu extension test`
 
 Destinations implemeted with this SDK:
 * [Mixpanel Destination](https://github.com/jitsucom/jitsu-mixpanel)
